- fix RockCleanup getting an exception if a Temporary BinaryFile was referenced by another table

+++ Rock/Jobs/RockCleanup.cs
@@ -190,8 +190,12 @@ namespace Rock.Jobs
             {
                 if ( binaryFile.ModifiedDateTime < RockDateTime.Now.AddDays( -1 ) )
                 {
-                    binaryFileService.Delete( binaryFile );
-                    binaryFileRockContext.SaveChanges();
+                    string errorMessage;
+                    if ( binaryFileService.CanDelete( binaryFile, out errorMessage ) )
+                    {
+                        binaryFileService.Delete( binaryFile );
+                        binaryFileRockContext.SaveChanges();
+                    }
                 }
             }
         }
